AI-Service/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0